# public/api-doc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Makkelijke Markt API",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/1.1.0/markt/": {
            "get": {
                "tags": [
                    "Markt"
                ],
                "summary": "Zoek door alle markten",
                "operationId": "MarktGetAll",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Markt"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/1.1.0/markt/{id}": {
            "get": {
                "tags": [
                    "Markt"
                ],
                "summary": "Vraag een markt op",
                "operationId": "MarktGetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Markt"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Markt"
                ],
                "summary": "Sla extra markt gegevens op die niet uit PerfectView komen",
                "operationId": "MarktPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "aantalKramen",
                                    "aantalMeter",
                                    "auditMax",
                                    "kiesJeKraamActief",
                                    "kiesJeKraamFase",
                                    "kiesJeKraamMededelingActief",
                                    "kiesJeKraamMededelingTekst",
                                    "kiesJeKraamMededelingTitel",
                                    "kiesJeKraamGeblokkeerdePlaatsen",
                                    "kiesJeKraamGeblokkeerdeData",
                                    "kiesJeKraamEmailKramenzetter",
                                    "makkelijkeMarktActief",
                                    "indelingstype",
                                    "marktDagenTekst",
                                    "indelingsTijdstipTekst",
                                    "telefoonNummerContact"
                                ],
                                "properties": {
                                    "aantalKramen": {
                                        "description": "Aantal kramen op de markt (capaciteit)",
                                        "type": "integer"
                                    },
                                    "aantalMeter": {
                                        "description": "Aantal meter op de markt (capaciteit)",
                                        "type": "integer"
                                    },
                                    "auditMax": {
                                        "description": "Aantal plaatsen op de audit lijst",
                                        "type": "integer"
                                    },
                                    "kiesJeKraamActief": {
                                        "type": "boolean"
                                    },
                                    "kiesJeKraamFase": {
                                        "type": "string"
                                    },
                                    "kiesJeKraamMededelingActief": {
                                        "type": "boolean"
                                    },
                                    "kiesJeKraamMededelingTekst": {
                                        "type": "string"
                                    },
                                    "kiesJeKraamMededelingTitel": {
                                        "type": "string"
                                    },
                                    "kiesJeKraamGeblokkeerdePlaatsen": {
                                        "type": "string"
                                    },
                                    "kiesJeKraamGeblokkeerdeData": {
                                        "type": "string"
                                    },
                                    "kiesJeKraamEmailKramenzetter": {
                                        "type": "string"
                                    },
                                    "makkelijkeMarktActief": {
                                        "type": "boolean"
                                    },
                                    "indelingstype": {
                                        "type": "string"
                                    },
                                    "marktDagenTekst": {
                                        "type": "string"
                                    },
                                    "indelingsTijdstipTekst": {
                                        "type": "string"
                                    },
                                    "telefoonNummerContact": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Markt"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [],
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Markt": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "14"
                    },
                    "afkorting": {
                        "type": "string",
                        "example": "NOM-M"
                    },
                    "naam": {
                        "type": "string",
                        "example": "Noordermarkt maandag"
                    },
                    "soort": {
                        "type": "string",
                        "example": "week"
                    },
                    "geoArea": {},
                    "marktDagen": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "ma",
                            "di"
                        ]
                    },
                    "standaardKraamAfmeting": {
                        "example": 3
                    },
                    "extraMetersMogelijk": {},
                    "aanwezigeOptiesResult": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": {
                            "3mKramen": true,
                            "4mKramen": false,
                            "extraMeters": true,
                            "elektra": true
                        }
                    },
                    "perfectViewNummer": {
                        "type": "integer",
                        "example": 40
                    },
                    "aantalKramen": {
                        "type": "integer"
                    },
                    "aantalMeter": {
                        "type": "integer"
                    },
                    "auditMax": {
                        "type": "integer",
                        "example": 10
                    },
                    "kiesJeKraamMededelingActief": {
                        "type": "boolean"
                    },
                    "kiesJeKraamMededelingTitel": {
                        "type": "string"
                    },
                    "kiesJeKraamMededelingTekst": {
                        "type": "string"
                    },
                    "kiesJeKraamActief": {
                        "type": "boolean"
                    },
                    "kiesJeKraamFase": {
                        "type": "string"
                    },
                    "kiesJeKraamGeblokkeerdePlaatsen": {
                        "type": "string"
                    },
                    "kiesJeKraamGeblokkeerdeData": {
                        "type": "string",
                        "example": "2019-12-25,2019-12-26"
                    },
                    "kiesJeKraamEmailKramenzetter": {
                        "type": "string",
                        "example": "test@domain.tld"
                    },
                    "marktDagenTekst": {
                        "type": "string"
                    },
                    "indelingsTijdstipTekst": {
                        "type": "string"
                    },
                    "telefoonNummerContact": {
                        "type": "string"
                    },
                    "makkelijkeMarktActief": {
                        "type": "boolean"
                    },
                    "indelingstype": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "MmAppKey",
                "in": "header"
            },
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}